+++ package.json
@@ -1,32 +1,34 @@
 {
   "name": "developers.sparkpost.com",
-  "description": "SparkPost Developer Hub",
+  "description": "SparkPost Developer Site",
   "version": "1.0.0",
   "author": "SparkPost <dev-relations@sparkpost.com>",
   "dependencies": {
     "babel-plugin-styled-components": "^1.5.1",
+    "extract-text-webpack-plugin": "^3.0.2",
     "font-awesome": "^4.7.0",
-    "gatsby": "^1.9.253",
-    "gatsby-link": "^1.6.32",
-    "gatsby-plugin-google-tagmanager": "^1.0.18",
-    "gatsby-plugin-lodash": "^1.0.11",
-    "gatsby-plugin-netlify-cms": "^1.0.12",
-    "gatsby-plugin-page-creator": "^1.0.1",
-    "gatsby-plugin-react-helmet": "^1.0.8",
-    "gatsby-plugin-react-next": "^1.0.11",
-    "gatsby-plugin-root-import": "^1.0.0",
-    "gatsby-plugin-styled-components": "^2.0.6",
+    "gatsby": "^2.0.0-beta.1",
+    "gatsby-plugin-google-tagmanager": "^2.0.0-beta.1",
+    "gatsby-plugin-lodash": "^3.0.1-beta.1",
+    "gatsby-plugin-netlify-cms": "^2.0.0-beta.1",
+    "gatsby-plugin-page-creator": "^2.0.0-beta.1",
+    "gatsby-plugin-react-helmet": "^3.0.0-beta.1",
+    "gatsby-plugin-styled-components": "^3.0.0-beta.1",
     "gatsby-plugin-webpack-bundle-analyzer": "^0.1.1",
-    "gatsby-source-filesystem": "^1.5.33",
+    "gatsby-source-filesystem": "^2.0.1-beta.1",
     "gatsby-source-github": "0.0.1",
-    "gatsby-source-wordpress": "^2.0.76",
-    "gatsby-transformer-json": "^1.0.16",
+    "gatsby-source-wordpress": "^3.0.0-beta.1",
+    "gatsby-transformer-json": "^2.1.1-beta.1",
     "hast-util-raw": "^2.0.2",
     "is-absolute-url": "^2.1.0",
     "lodash": "^4.17.10",
     "minim": "^0.20.5",
     "minim-parse-result": "^0.10.0",
+    "netlify-cms": "^1.9.2",
     "polished": "^1.9.2",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
+    "react-helmet": "^5.2.0",
     "react-instantsearch": "^5.1.0",
     "react-sticky": "^6.0.2",
     "rehype-highlight": "^2.1.0",
@@ -45,7 +47,7 @@
   "scripts": {
     "bootstrap": "npm install && cd plugins/gatsby-transformer-api-elements && npm install",
     "develop": "source .env && gatsby develop",
-    "docs": "NODE_ENV=docs && gatsby develop",
+    "docs": "ACTIVE_ENV=docs gatsby develop",
     "build": "gatsby build",
     "format": "prettier --trailing-comma es5 --no-semi --single-quote --write \"src/**/*.js\" \"content/**/*.js\"",
     "precommit": "npm run format"
