+++ src/components/Grid/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { mediaQuery } from '../../utils/breakpoint'
+import { mediaQuery } from 'utils/breakpoint'
 
 import Row from './Row'
 import Column from './Column'
@@ -11,10 +11,9 @@ const Container = styled.div`
   padding-left: 15px;
   padding-right: 15px;
 
-  ${mediaQuery('sm', 'width: 750px;')} ${mediaQuery(
-      'md',
-      'width: 970px;'
-    )} ${mediaQuery('lg', 'width: 1170px;')};
+  ${mediaQuery('sm', 'width: 750px;')}
+  ${mediaQuery('md', 'width: 970px;')}
+  ${mediaQuery('lg', 'width: 1170px;')}
 `
 
 export { Container, Row, Column }
